Prefer reading query from stdin if available
A recent change made it so not providing a query causes everything to be
exported. However, that made it so providing a query on stdin required
the option `vast.export.read` to be set to `"-"`, which was far from
intuitive. This change streamlines the option, giving the user an error
if conflicting ways to read the query exist and preferring stdin over
exporting everything if a query is provided on stdin but not on the
command line.