Move one-off up the call stack
Turns out that removing the uhash-inspired seeded_hash function object
had the side-effect of enabling the fast path for all objects. For
addresses, this means we're not going through hash_append anymore, so
our one-off is dead code.

Until we have versioned the persistent state of the sketch structures,
this hack has to remain in there.