Fix deadlock for queries with no candidate partitions
With the introduction of the query backlog we changed how we handle the
available query workers. This introduced a subtle bug: Queries from
anonymous senders and those that did not qualify for any candidate
partitions from the meta index did not return the query worker, causing
it to be unavailable permanently. This meant that VAST was unable to
export data after handling exactly as many queries of that kind as there
are query workers available, resulting in a deadlock.