Port integration tests to bats
Moving away from the homegrown integration testing tool to a test
framework that works directly in the shell environment offers a
big increase of flexibility at a rather modest cost in verbosity.

This solves 2 long-standing pain points:
- it is possible to test concurrent client processes. See
  `database.bats:parallel clients` for an example.
- Disk- and IPC Message format stability can be tested by running
  specific older versions of VAST via docker or nix directly inside
  of a test.

The overhead of the test runner itself is negligible compared to
the commands that regularly executed as part of the tests. I ported
the "Server Zeek conn log" integration test (now called "batch sizes")
over, and both frameworks indeed show comparable execution times.

There is one drawback that needs to be mentioned:
`bats` has no functionality to update expectation values like the
integration frameworks `--update` flag. It may be possible to
retrofit this features with a custom wrapper for the `assert_output`
function.